Revert "Squashed commit of the following:"
This reverts commit c51022993b032c7b4c3618dde7ca10435aa884f9.

+++ src/upgrades/1.7.1/notification-settings.js
@@ -8,38 +8,23 @@ module.exports = {
 	timestamp: Date.UTC(2017, 10, 15),
 	method: async function () {
 		const { progress } = this;
-		progress.total = await db.sortedSetCard('users:joindate');
+
 		await batch.processSortedSet('users:joindate', async (uids) => {
-
-			const userSettings = await db.getObjectsFields(
-				uids.map(uid => `user:${uid}:settings`),
-				['sendChatNotifications', 'sendPostNotifications'],
-			);
-
-			const bulkSet = [];
-			userSettings.forEach((settings, index) => {
-				const set = {};
-				if (settings) {
+			await Promise.all(uids.map(async (uid) => {
+				progress.incr();
+				const userSettings = await db.getObjectFields(`user:${uid}:settings`, ['sendChatNotifications', 'sendPostNotifications']);
+				if (userSettings) {
 					if (parseInt(userSettings.sendChatNotifications, 10) === 1) {
-						set['notificationType_new-chat'] = 'notificationemail';
+						await db.setObjectField(`user:${uid}:settings`, 'notificationType_new-chat', 'notificationemail');
 					}
 					if (parseInt(userSettings.sendPostNotifications, 10) === 1) {
-						set['notificationType_new-reply'] = 'notificationemail';
-					}
-					if (Object.keys(set).length) {
-						bulkSet.push([`user:${uids[index]}:settings`, set]);
+						await db.setObjectField(`user:${uid}:settings`, 'notificationType_new-reply', 'notificationemail');
 					}
 				}
-			});
-			await db.setObjectBulk(bulkSet);
-
-			await db.deleteObjectFields(
-				uids.map(uid => `user:${uid}:settings`),
-				['sendChatNotifications', 'sendPostNotifications'],
-			);
-
-			progress.incr(uids.length);
+				await db.deleteObjectFields(`user:${uid}:settings`, ['sendChatNotifications', 'sendPostNotifications']);
+			}));
 		}, {
+			progress: progress,
 			batch: 500,
 		});
 	},
